refactor: todoReducer 함수 분리
이때 export default 함수이름을 사용하면 안됨.
- export default function 으로 사용시 expression(표현식)으로 표현해야하고,
- arrow function을 사용하고 싶다면 declaration해주고 아래에 export default 로 빼줘야한다.
