+++ src/TodoApp.js
@@ -5,59 +5,13 @@ import InputBar from "./Header/InputBar";
 import Loader from "./Loader";
 import ResultBar from "./Header/ResultBar";
 import { StateProvider } from "./TodoState";
+import todosReducer from "./todoReducer";
 
 export default function TodoApp() {
   const [todos, setTodos] = useState([]);
-  const [newTodo, setNewTodo] = useState("");
 
   const initialState = {
     todos
-  };
-
-  const changeInput = (newTodo) => {
-    setNewTodo(newTodo);
-  }
-
-  const todosReducer = (state, action) => {
-    switch (action.type) {
-      case "ADD_TODO":
-        const newTodoObj = {
-          title: action.title,
-          id: action.id,
-          status: "todo"
-        };
-        return {
-          todos: [...state.todos, newTodoObj],
-        };
-      case "DELETE_TODO":
-        const remaindedTodos = state.todos.filter(
-          todo => todo.id !== action.id
-        );
-        return {
-          todos: remaindedTodos
-        };
-
-      case "CHANGE_TODO_STATUS":
-        const index = state.todos.findIndex(todo => todo.id === action.id);
-        const selected = state.todos[index];
-        const nextTodos = [...state.todos];
-
-        const statusToggle = status => {
-          return status === "todo" ? "done" : "todo";
-        };
-
-        nextTodos[index] = {
-          ...selected,
-          status: statusToggle(selected.status)
-        };
-
-        return {
-          todos: nextTodos
-        };
-
-      default:
-        return state;
-    }
   };
 
   useEffect(() => {
@@ -79,7 +33,7 @@ export default function TodoApp() {
   return initialState.todos.length !== 0 ? (
     <StateProvider initialState={initialState} reducer={todosReducer}>
       <ResultBar />
-      <InputBar  newTodo={newTodo} changeInput={changeInput}/>
+      <InputBar />
       <TodoList />
     </StateProvider>
   ) : (
